feature changeorderstatus

+++ WebView/Areas/BanHangOnline/HoangDTO/Param/ThanhToanParam.cs
@@ -3,6 +3,6 @@
     public class ThanhToanParam
     {
         public int IdMaGiamGia { get; set; } = 0;
-        public double PhiVanChuyen { get; set; } = 0;
+        public decimal PhiVanChuyen { get; set; } = 0;
     }
 }
